@@ -26,5 +26,5 @@ import (
 var emojis = []string{":beer:", ":pizza:", ":taco:"}
 
 func init() {
-	rand.Seed(time.Now().UTC().UnixNano())
+	rand.New(rand.NewSource(time.Now().UTC().UnixNano()))
 }
